% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinations.R
\name{gen_combinations}
\alias{gen_combinations}
\title{Generate levels of all combinations}
\usage{
gen_combinations(nms, collapse = "/")
}
\arguments{
\item{nms}{Character vector to generate combinations}

\item{collapse}{Character string to separate the combinations}
}
\description{
Generate levels of all combinations
}
