% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/or.R
\name{or_plot}
\alias{or_plot}
\title{Odds Ratio Plot Helper}
\usage{
or_plot(
  mod,
  x_pos = 2,
  x_pos_sh = rep(0.5, 2),
  breaks = c(0.5, 1, 1.5),
  nms = c(),
  order = NULL
)
}
\arguments{
\item{mod}{Linear model output}

\item{x_pos}{X axis position of ORs column}

\item{x_pos_sh}{X axis position start and end of underline in OR columns}

\item{breaks}{X axis breaks}

\item{nms}{Named vector for term labels in model}

\item{order}{Vector for order of terms}
}
\description{
Odds Ratio Plot Helper
}
