% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{rbind_list_base}
\alias{rbind_list_base}
\title{Bind list of data frames together}
\usage{
rbind_list_base(x)
}
\arguments{
\item{x}{List of data.frames to bind together}
}
\description{
Bind list of data frames together
}
\examples{
{

df1 <- data.frame("a" = 1, "b" = 2)
df2 <- data.frame("a" = 2, "b" = 4)
l <- list(df1, df2)
df3 <- rbind_list_base(l)
df3

}
}
